Adds arxiv.db.init()
Breaking change:
arxiv.db.models.configure_db moved to arxiv.db.configure_db
