+++ arxiv/auth/legacy/tests/test_endorsements.py
@@ -11,6 +11,7 @@ from flask import Flask
 from mimesis import Person, Internet, Datetime
 from sqlalchemy import insert
 
+import arxiv.db
 from arxiv.taxonomy import definitions
 from arxiv.config import Settings
 from arxiv.db import models, transaction
@@ -33,7 +34,8 @@ class TestEndorsement(TestCase):
                         CLASSIC_DB_URI=self.app.config['CLASSIC_DATABASE_URI'],
                         LATEXML_DB_URI=None)
 
-        engine, _ = models.configure_db(settings)
+        arxiv.db.init(settings)
+        util.create_all(arxiv.db._classic_engine)
 
         self.default_tracking_data = {
             'remote_addr': '0.0.0.0',
@@ -42,7 +44,6 @@ class TestEndorsement(TestCase):
         }
 
         with self.app.app_context():
-            util.create_all(engine)
             with transaction() as session:
                 person = Person('en')
                 net = Internet()
@@ -165,7 +166,8 @@ class TestAutoEndorsement(TestCase):
                         CLASSIC_DB_URI=self.app.config['CLASSIC_DATABASE_URI'],
                         LATEXML_DB_URI=None)
 
-        engine, _ = models.configure_db(settings)
+        arxiv.db.init(settings)
+        util.create_all(arxiv.db._classic_engine)
 
         self.default_tracking_data = {
             'remote_addr': '0.0.0.0',
@@ -174,7 +176,6 @@ class TestAutoEndorsement(TestCase):
         }
 
         with self.app.app_context():
-            util.create_all(engine)
             with transaction() as session:
                 person = Person('en')
                 net = Internet()
