+++ arxiv/auth/legacy/tests/util.py
@@ -11,9 +11,9 @@ from flask import Flask
 from pytz import UTC
 from sqlalchemy import select
 
+import arxiv.db
 from arxiv.config import Settings
-from arxiv.db import transaction, models, Session
-from arxiv.db.models import configure_db
+from arxiv.db import transaction, models, Session, configure_db
 from .. import util
 from ..passwords import hash_password
 
@@ -29,17 +29,17 @@ def temporary_db(db_uri: str, create: bool = True, drop: bool = True):
                         CLASSIC_DB_URI=db_uri,
                         LATEXML_DB_URI=None)
 
-    engine, _ = configure_db (settings)
+    arxiv.db.init(settings)
+    if create:
+        util.create_all(arxiv.db._classic_engine)
 
     with app.app_context():
-        if create:
-            util.create_all(engine)
         try:
             with transaction() as session:
                 yield session
         finally:
             if drop:
-                util.drop_all(engine)
+                util.drop_all(arxiv.db._classic_engine)
 
 
 class SetUpUserMixin(TestCase):
@@ -62,7 +62,7 @@ class SetUpUserMixin(TestCase):
             CLASSIC_DB_URI=self.db_uri,
             LATEXML_DB_URI=None)
 
-        self.engine, _ = models.configure_db(settings)
+        self.engine, _ = arxiv.db.configure_db(settings)
 
         with self.app.app_context():
                 # Insert tapir policy classes
