+++ arxiv/db/models.py
@@ -5,7 +5,6 @@ This file was generated by using sqlacodgen
 
 from typing import Optional, Literal, Any, Tuple, List
 import re
-import os
 import hashlib
 import datetime as dt
 from datetime import datetime, date
@@ -30,8 +29,7 @@ from sqlalchemy import (
     Text, 
     Table,
     Enum,
-    text, 
-    create_engine,
+    text,
     Engine
 )
 from sqlalchemy.schema import FetchedValue
@@ -41,73 +39,15 @@ from sqlalchemy.orm import (
     relationship
 )
 
-from ..config import Settings, settings
+from ..config import settings
 from . import Base, LaTeXMLBase, metadata, \
-    session_factory
+    session_factory, _classic_engine, _latexml_engine
 
 from .types import intpk
 
 tb_secret = settings.TRACKBACK_SECRET
 tz = gettz(settings.ARXIV_BUSINESS_TZ)
 
-def configure_db (base_settings: Settings) -> Tuple[Engine, Optional[Engine]]:
-    if 'sqlite' in base_settings.CLASSIC_DB_URI:
-        engine = create_engine(base_settings.CLASSIC_DB_URI)
-        if base_settings.LATEXML_DB_URI:
-            latexml_engine = create_engine(base_settings.LATEXML_DB_URI)
-        else:
-            latexml_engine = None
-    else:
-        engine = create_engine(base_settings.CLASSIC_DB_URI,
-                        echo=base_settings.ECHO_SQL,
-                        isolation_level=base_settings.CLASSIC_DB_TRANSACTION_ISOLATION_LEVEL,
-                        pool_recycle=600,
-                        max_overflow=(base_settings.REQUEST_CONCURRENCY - 5), # max overflow is how many + base pool size, which is 5 by default
-                        pool_pre_ping=base_settings.POOL_PRE_PING)
-        if base_settings.LATEXML_DB_URI:
-            stmt_timeout: int = max(base_settings.LATEXML_DB_QUERY_TIMEOUT, 1)
-            latexml_engine = create_engine(base_settings.LATEXML_DB_URI,
-                                    connect_args={"options": f"-c statement_timeout={stmt_timeout}s"},
-                                    echo=base_settings.ECHO_SQL,
-                                    isolation_level=base_settings.LATEXML_DB_TRANSACTION_ISOLATION_LEVEL,
-                                    pool_recycle=600,
-                                    max_overflow=(base_settings.REQUEST_CONCURRENCY - 5),
-                                    pool_pre_ping=base_settings.POOL_PRE_PING)
-        else:
-            latexml_engine = None
-
-    return configure_db_engine(engine, latexml_engine)
-
-
-def configure_db_engine(classic_engine: Engine, latexml_engine: Optional[Engine]) -> Tuple[Engine, Optional[Engine]]:
-    session_factory.configure(binds={
-        Base: classic_engine,
-        LaTeXMLBase: (latexml_engine if latexml_engine else classic_engine),
-        t_arXiv_stats_hourly: classic_engine,
-        t_arXiv_admin_state: classic_engine,
-        t_arXiv_bad_pw: classic_engine,
-        t_arXiv_black_email: classic_engine,
-        t_arXiv_block_email: classic_engine,
-        t_arXiv_bogus_subject_class: classic_engine,
-        t_arXiv_duplicates: classic_engine,
-        t_arXiv_in_category: classic_engine,
-        t_arXiv_moderators: classic_engine,
-        t_arXiv_ownership_requests_papers: classic_engine,
-        t_arXiv_refresh_list: classic_engine,
-        t_arXiv_updates_tmp: classic_engine,
-        t_arXiv_white_email: classic_engine,
-        t_arXiv_xml_notifications: classic_engine,
-        t_demographics_backup: classic_engine,
-        t_tapir_email_change_tokens_used: classic_engine,
-        t_tapir_email_tokens_used: classic_engine,
-        t_tapir_error_log: classic_engine,
-        t_tapir_no_cookies: classic_engine,
-        t_tapir_periodic_tasks_log: classic_engine,
-        t_tapir_periodic_tasks_log: classic_engine,
-        t_tapir_permanent_tokens_used: classic_engine,
-        t_tapir_save_post_variables: classic_engine
-    })
-    return classic_engine, latexml_engine
 
 class MemberInstitution(Base):
     __tablename__ = 'Subscription_UniversalInstitution'
@@ -2127,3 +2067,35 @@ class DBLaTeXMLFeedback (LaTeXMLBase):
     selected_html: Mapped[Optional[str]]
     initiation_mode: Mapped[Optional[str]]
 
+
+def configure_db_engine(classic_engine: Engine, latexml_engine: Optional[Engine]) -> Tuple[Engine, Optional[Engine]]:
+    session_factory.configure(binds={
+        Base: classic_engine,
+        LaTeXMLBase: (latexml_engine if latexml_engine else classic_engine),
+        t_arXiv_admin_state: classic_engine,
+        t_arXiv_bad_pw: classic_engine,
+        t_arXiv_black_email: classic_engine,
+        t_arXiv_block_email: classic_engine,
+        t_arXiv_bogus_subject_class: classic_engine,
+        t_arXiv_duplicates: classic_engine,
+        t_arXiv_in_category: classic_engine,
+        t_arXiv_moderators: classic_engine,
+        t_arXiv_ownership_requests_papers: classic_engine,
+        t_arXiv_refresh_list: classic_engine,
+        t_arXiv_updates_tmp: classic_engine,
+        t_arXiv_white_email: classic_engine,
+        t_arXiv_xml_notifications: classic_engine,
+        t_demographics_backup: classic_engine,
+        t_tapir_email_change_tokens_used: classic_engine,
+        t_tapir_email_tokens_used: classic_engine,
+        t_tapir_error_log: classic_engine,
+        t_tapir_no_cookies: classic_engine,
+        t_tapir_periodic_tasks_log: classic_engine,
+        t_tapir_periodic_tasks_log: classic_engine,
+        t_tapir_permanent_tokens_used: classic_engine,
+        t_tapir_save_post_variables: classic_engine
+    })
+    return classic_engine, latexml_engine
+
+# Configure the models from env vars at package load time
+configure_db_engine(_classic_engine, _latexml_engine)
